docs: cleanup safety with docstrings, consolidate code, and general housekeeping
