+++ Cargo.toml
@@ -4,7 +4,7 @@ version = "0.14.0"
 edition = "2021"
 rust-version = "1.75.0"
 authors = ["A. Coull"]
-description = "PRISM-Q — Performance Rust Interoperable Simulator for Quantum"
+description = "PRISM-Q: Performance Rust Interoperable Simulator for Quantum"
 license = "MIT OR Apache-2.0"
 readme = "README.md"
 documentation = "https://docs.rs/prism-q"
