+++ deny.toml
@@ -3,7 +3,7 @@ all-features = true
 
 [advisories]
 ignore = [
-    # paste — unmaintained, transitive dep from faer/gemm. Revisit when upstream moves to pastey.
+    # paste: unmaintained, transitive dep from faer/gemm. Revisit when upstream moves to pastey.
     { id = "RUSTSEC-2024-0436", reason = "transitive dep from faer, revisit when deps migrate to pastey" },
 ]
